Changed all port to short
git-svn-id: https://dev.chiasoft.net:8443/svn/svn_vana/trunk@254 23f3cdc0-e6d1-42be-a0c1-51f4125517bb

+++ src/ChannelServer/PlayerPacket.cpp
@@ -223,7 +223,7 @@ void PlayerPacket::newSkin(Player* player){
 	packet.packetSend(player);
 }
 
-void PlayerPacket::changeChannel(Player *player, char *ip, int port) {
+void PlayerPacket::changeChannel(Player *player, char *ip, short port) {
 	Packet packet = Packet();
 	packet.addHeader(SEND_CHANGE_CHANNEL);
 	packet.addByte(1);
