+++ src/Common/Connection/Acceptor.h
@@ -23,7 +23,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 
 class Acceptor: public Selector::SelectHandler {
 public:
-	Acceptor (int port, AbstractPlayerFactory* apf);
+	Acceptor (short port, AbstractPlayerFactory* apf);
 	virtual void handle (int socket);
 protected:
 	AbstractPlayerFactory* abstractPlayerFactory;
